+++ src/main/java/com/somemore/volunteerapply/usecase/ApplyVolunteerApplyUseCase.java
@@ -1,6 +1,6 @@
 package com.somemore.volunteerapply.usecase;
 
-import com.somemore.volunteerapply.dto.VolunteerApplyCreateRequestDto;
+import com.somemore.volunteerapply.dto.request.VolunteerApplyCreateRequestDto;
 import java.util.UUID;
 
 public interface ApplyVolunteerApplyUseCase {
